# swagger-doc.json
{"openapi":"3.0.0","paths":{"/api/health/hc":{"get":{"operationId":"HealthController_check","summary":"Check all services are up or down","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}}}}},"/api/v1/users":{"post":{"operationId":"UsersController_create","summary":"Create a user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"statusCode":{"type":"number","default":200},"data":{"type":"object","$ref":"#/components/schemas/CreateUserDto"}}}}}}},"tags":["Users"]},"get":{"operationId":"UsersController_findAll","summary":"Get list user","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"statusCode":{"type":"number","default":200},"data":{"type":"array","items":{"type":"object","$ref":"#/components/schemas/CreateUserDto"}}}}}}}},"tags":["Users"]}},"/api/v1/users/{username}":{"get":{"operationId":"UsersController_findOne","summary":"Get user detail","parameters":[{"name":"username","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"statusCode":{"type":"number","default":200},"data":{"type":"object","$ref":"#/components/schemas/CreateUserDto"}}}}}}},"tags":["Users"]},"delete":{"operationId":"UsersController_remove","summary":"Delete a user","parameters":[{"name":"username","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"statusCode":{"type":"number","default":200},"message":{"type":"string","example":"success"}}}}}}},"tags":["Users"]}}},"info":{"title":"Stock","description":"Swagger document for Stock APIs","version":"1.0","contact":{}},"tags":[{"name":"Stock","description":""}],"servers":[{"url":"http://[::1]:5001"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"age":{"type":"number"},"address":{"type":"string"},"avatar":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]}}}}